Hengle/other permission

+++ src/CrectoAdmin/resource.cr
@@ -43,7 +43,8 @@ def self.admin_resource(model : Crecto::Model.class, repo, **opts)
 
   # Index
   get "/admin/#{model.table_name}" do |ctx|
-    access = CrectoAdmin.model_access(ctx, resource)
+    user = CrectoAdmin.current_user(ctx)
+    access = CrectoAdmin.check_access(user, resource)
     next if access[0].nil? || access[1].empty?
     model_query = access[0].as(Crecto::Repo::Query)
     collection_attributes = resource[:collection_attributes].select { |a| access[1].includes? a }
@@ -51,12 +52,14 @@ def self.admin_resource(model : Crecto::Model.class, repo, **opts)
     query = model_query.limit(per_page).offset(offset)
     data = repo.all(model, query)
     count = repo.aggregate(model, :count, resource[:model].primary_key_field_symbol, model_query).as(Int64)
+    form_attributes = CrectoAdmin.check_create(user, resource, access[1])
     ecr("index")
   end
 
   # Search
   get "/admin/#{model.table_name}/search" do |ctx|
-    access = CrectoAdmin.model_access(ctx, resource)
+    user = CrectoAdmin.current_user(ctx)
+    access = CrectoAdmin.check_access(user, resource)
     next if access[0].nil? || access[1].empty?
     model_query = access[0].as(Crecto::Repo::Query)
     collection_attributes = resource[:collection_attributes].select { |a| access[1].includes? a }
@@ -64,26 +67,32 @@ def self.admin_resource(model : Crecto::Model.class, repo, **opts)
     search_attributes.delete(model.primary_key_field_symbol)
     search_attributes.unshift(model.primary_key_field_symbol)
     offset = ctx.params.query["offset"]? ? ctx.params.query["offset"].to_i : 0
-    search_string = search_attributes.map { |sa| "#{CrectoAdmin.field_cast(sa, repo)} LIKE ?" }.join(" OR ")
+    search_string = "(" + search_attributes.map { |sa| "#{CrectoAdmin.field_cast(sa, repo)} LIKE ?" }.join(" OR ") + ")"
     search_params = (1..search_attributes.size).map { |x| "%#{CrectoAdmin.search_param(ctx)}%" }
     query = model_query
       .limit(per_page)
       .where(search_string, search_params)
       .offset(offset)
     data = repo.all(model, query)
-    count = repo.aggregate(model, :count, resource[:model].primary_key_field_symbol, model_query).as(Int64)
+    count = repo.aggregate(model, :count, resource[:model].primary_key_field_symbol, model_query.where(search_string, search_params)).as(Int64)
+    form_attributes = CrectoAdmin.check_create(user, resource, access[1])
     ecr("index")
   end
 
   # New form
   get "/admin/#{model.table_name}/new" do |ctx|
+    user = CrectoAdmin.current_user(ctx)
+    access = CrectoAdmin.check_access(user, resource)
+    next if access[0].nil? || access[1].empty?
     item = model.new
+    form_attributes = CrectoAdmin.check_create(user, resource, access[1])
     ecr("new")
   end
 
   # View
   get "/admin/#{model.table_name}/:id" do |ctx|
-    access = CrectoAdmin.model_access(ctx, resource)
+    user = CrectoAdmin.current_user(ctx)
+    access = CrectoAdmin.check_access(user, resource)
     next if access[0].nil? || access[1].empty?
     model_query = access[0].as(Crecto::Repo::Query)
     query = model_query.where(resource[:model].primary_key_field_symbol, ctx.params.url["id"])
@@ -91,14 +100,20 @@ def self.admin_resource(model : Crecto::Model.class, repo, **opts)
     next if data.empty?
     item = data.first
     model_attributes = access[1]
+    form_attributes = CrectoAdmin.check_edit(user, resource, item, access[1])
+    can_delete = CrectoAdmin.check_delete(user, resource, item, form_attributes)
     ecr("show")
   end
 
   # Update
   put "/admin/#{model.table_name}/:pid_id" do |ctx|
+    user = CrectoAdmin.current_user(ctx)
+    access = CrectoAdmin.check_access(user, resource)
+    next if access[0].nil? || access[1].empty?
     item = repo.get!(model, ctx.params.url["pid_id"])
+    form_attributes = CrectoAdmin.check_edit(user, resource, item, access[1])
     query_hash = ctx.params.body.to_h
-    resource[:form_attributes].each do |attr|
+    form_attributes.each do |attr|
       next if attr.is_a? Symbol
       attr = attr.as(Tuple(Symbol, String) | Tuple(Symbol, String, Array(String) | String))
       if attr[1] == "bool"
@@ -127,13 +142,20 @@ def self.admin_resource(model : Crecto::Model.class, repo, **opts)
 
   # Edit form
   get "/admin/#{model.table_name}/:id/edit" do |ctx|
-    item = repo.get!(model, ctx.params.url["id"])
+    user = CrectoAdmin.current_user(ctx)
+    access = CrectoAdmin.check_access(user, resource)
+    next if access[0].nil? || access[1].empty?
+    model_query = access[0].as(Crecto::Repo::Query)
+    query = model_query.where(resource[:model].primary_key_field_symbol, ctx.params.url["id"])
+    data = repo.all(model, query).not_nil!
+    next if data.empty?
+    item = data.first
+    form_attributes = CrectoAdmin.check_edit(user, resource, item, access[1])
     ecr("edit")
   end
 
   # Create
   post "/admin/#{model.table_name}" do |ctx|
-    puts "create"
     item = model.new
     query_hash = ctx.params.body.to_h
     resource[:form_attributes].each do |attr|
@@ -165,7 +187,13 @@ def self.admin_resource(model : Crecto::Model.class, repo, **opts)
 
   # Delete
   get "/admin/#{model.table_name}/:id/delete" do |ctx|
+    user = CrectoAdmin.current_user(ctx)
+    access = CrectoAdmin.check_access(user, resource)
+    next if access[0].nil? || access[1].empty?
     item = repo.get!(model, ctx.params.url["id"])
+    form_attributes = CrectoAdmin.check_edit(user, resource, item, access[1])
+    can_delete = CrectoAdmin.check_delete(user, resource, item, form_attributes)
+    next unless can_delete
     changeset = repo.delete(item)
 
     if changeset.errors.any?
